chore : add eventTicketPrice entity chore : refactor price and quantity data types in event ticket prices table in the migration script

+++ src/main/resources/db/migration/V1__innit.sql
@@ -72,8 +72,8 @@ CREATE TABLE event_ticket_prices (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     event_id BIGINT NOT NULL,
     ticket_type_id BIGINT NOT NULL,
-    price BIGINT NOT NULL,
-    quantity BIGINT NOT NULL,
+    price INT NOT NULL,
+    quantity INT NOT NULL,
 
     FOREIGN KEY (event_id) REFERENCES  events(id),
     FOREIGN KEY (ticket_type_id) REFERENCES ticket_types(id)
